ReactionUpdate ctrl and other updates all implemented. Tested all situations/cases, everything works

+++ backend/models/reactions.js
@@ -19,7 +19,7 @@ const Reactions = sequelize.define('Reactions', {
     },
 
     ReactionType:{
-        type: DataTypes.TINYINT
+        type: DataTypes.SMALLINT
     }
 },
 {
